requirements_test.txt
django-imperavi
django-tinymce
pytz
Pillow
webtest
django-webtest
